An other Fix for LastUpdate...

+++ proxytest/trunk/CAUtil.hpp
@@ -159,6 +159,7 @@ inline bool isGreater64(UINT64& op1,UINT64& op2)
 		#endif
 	}
 */
+
 inline void print64(UINT8* buff,UINT64 num)
 	{
 		#ifdef HAVE_NATIVE_UINT64
@@ -168,7 +169,7 @@ inline void print64(UINT8* buff,UINT64 num)
 					buff[1]=0;
 					return;
 				}
-			UINT64 mask=(UINT64)10000000000000000000UL;
+			UINT64 mask=10000000000000000000ULL;
 			UINT digit;
 			UINT32 index=0;
 			bool bprintZero=false;
